requirements.txt
sounddevice
numpy
pycairo
PyGObject
pynput
faster-whisper
